polish redis identity directory API a bit
